github-fetch-reviews: Prevent re-fetching already downloaded reviews
I've been getting

    OSError: [Errno 12] Cannot allocate memory

errors in relation to urllib3, and rather than look into why that's
happening or how to fix it, I decided to just keep running the script
until it gets all the data. But when re-running the script, I don't want
it to re-fetch data I already have.