github-fetch-reviews: Get owner-repo from base repository
Previously, the reviews were requested from the 'head' repository. This
didn't matter when the 'head' branch was in the same repository as the
'base', but GitHub forks will have a different owner-repo value than
what we need to fetch the review data. Use 'base' instead of 'head' in
order to get the correct owner-repo.