Get 'days until merge' graph
Add a new script to get the number of days that pull requests stay open
until they're merged.

Similar to the time until first review, but here approximating for how
much back and forth there could be on a change request.

Copied much of the code from the other scripts.